Move `Player` to its own module
